The Dialog Box is Learned Using the Getx StateManagement
